% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation_fxns.R
\name{fill_tag_id}
\alias{fill_tag_id}
\title{Fill in full Vue Tag ID}
\usage{
fill_tag_id(dat, tags = NA, db = db)
}
\arguments{
\item{dat}{A dataframe of \code{acoustic_animals} data; typically from the output of \code{prep_otn_tagging}.}

\item{tags}{A dataframe of \code{metadata_acoustictags} data that is not yet in the database to supplement tag lookups; typically the output of \code{prep_tag_sheet}.}

\item{db}{Name of SOMNI database connection object in R workspace. Defaults to "db".}
}
\value{
A dataframe that matches the input \code{dat} dataframe, but with the \code{vue_tag_id} column populated.
}
\description{
If you have empty \code{vue_tag_id} values in your
\code{acoustic_animals} dataframe, you can use \code{fill_tag_id} to
find and populate \code{vue_tag_id} using the \code{vue_id} and
\code{tag_serial} columns. Any existing \code{vue_tag_id} values are
not modified.
}
\details{
Each complete tag ID (e.g., 'A69-9001-3945') has a unique
Vue ID and tag serial. Provided you have the Vue ID and serial
number, in addition to the original tag purchase sheets, you
can easily look up the complete tag ID.
}
\examples{
tm <- readxl::read_excel("OTN_tagging_sheet.xslx", sheet = "Tag Metadata")
prepped_otn <- prep_otn_tagging(tm, db = db)
fill_tag_id(prepped_otn$acoustic_animals, db = db)
}
