% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_prep_fxns.R
\name{prep_otn_deployment}
\alias{prep_otn_deployment}
\title{Prepare OTN instrument metadata sheet for SOMNI db}
\usage{
prep_otn_deployment(dat, db = db)
}
\arguments{
\item{dat}{A dataframe containing OTN tagging metadata, including the template header with the OTN logo, template version number, and instructions.}

\item{db}{Name of SOMNI database connection object in R workspace. Defaults to "db".}
}
\value{
A list containing up to 11 dataframes, either corresponding to their respective table in SOMNI db or containing errors that need to be fixed.
}
\description{
Prepare OTN instrument metadata sheet for SOMNI db
}
