% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_prep_fxns.R
\name{prep_otn_tagging}
\alias{prep_otn_tagging}
\title{Prepare OTN tagging metadata sheets for SOMNI db}
\usage{
prep_otn_tagging(dat)
}
\arguments{
\item{dat}{A dataframe containing OTN tagging metadata, including the template header with the OTN logo and instructions.}
}
\value{
A list containing two dataframes, each corresponding to the respective table in SOMNI db: \code{metadata_animals} and \code{acoustic_animals}.
}
\description{
Prepare OTN tagging metadata sheets for SOMNI db
}
\examples{
tm <- readxl::read_excel("OTN_tagging_sheet.xslx", sheet = "Tag Metadata")
prep_otn_tagging(tm)
}
