% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_prep_fxns.R
\name{prep_tag_sheet}
\alias{prep_tag_sheet}
\title{Prepare Vemco/Innovasea tag sheets for SOMNI db}
\usage{
prep_tag_sheet(tags)
}
\arguments{
\item{tags}{Dataframe or tibble of tag sheet data. Multiple tag sheets can be merged into one dataframe.}
}
\value{
A dataframe with 45 columns, matching the table structure in the SOMNI db table \code{metadata_acoustictags}.
}
\description{
This function takes Vemco/Innovasea sales order tag sheets,
performs simple data validation, selects necessary columns for SOMNI db,
and ensures all columns are coerced to the correct data type.
}
\examples{
TagSheet <- read.csv("TagSheet.csv")
prep_tag_sheet(TagSheet)
}
