% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation_fxns.R
\name{validate_acoustic_animals}
\alias{validate_acoustic_animals}
\title{Validate acoustic_animals data}
\usage{
validate_acoustic_animals(dat, tags = NA, db = db)
}
\arguments{
\item{dat}{A dataframe of \code{acoustic_animals} data; typically from the output of \code{prep_otn_tagging}.}

\item{tags}{A dataframe of \code{metadata_acoustictags} data that is not yet in the database to supplement error scanning; typically the output of \code{prep_tag_sheet}.}

\item{db}{Name of SOMNI database connection object in R workspace. Defaults to "db".}
}
\value{
A tibble containing all errors flagged.
}
\description{
\code{validate_acoustic_animals} scans the output of
\code{prep_otn_tagging} and flags any tag-related errors it finds.
}
\details{
\code{prep_otn_tagging} prepares OTN tag sheet data for
ingestion into the SOMNI database. The most important output of that
is the \code{acoustic_animals} table, which links each unique animal
to a unique acoustic tag. This table is ultimately used to link
receiver detections to an animal in the database. As such, it's
important to ensure the integrity of your tag metadata!

This function validates tag metadata in three ways: first, it ensures
the required tag parameters (either tag serial number, \code{tag_serial},
OR the full Vue tag ID, \code{vue_tag_id}) are present in the data.

Second, it scans tag sheets to ensure all tags have a match in the
database. This is helpful for flagging any serial numbers or IDs that
may have typos in them. In order to take full advantage of this, a
database connection must be provided in the \code{db} arg of the function.
If you have any tag sheets that you have processed with \code{prep_tag_sheet}
but not yet imported into the database, they can be provided as a
dataframe via the \code{tags} arg.

Finally, if provided an active database connection via \code{db}, this
function checks if any tags you are validating are already paired with
another animal in the database. This means the tag ID you have provided
is either a typo OR comes from an animal that has been recaptured,
and the tag is now being re-used. Remember, a tag cannot be in two
animals at once! This will only cause errors in your detection data if
recaptures are not properly updated in the database.
}
\examples{
\dontrun{
tm <- readxl::read_excel("OTN_tagging_sheet.xslx", sheet = "Tag Metadata")
prepped_otn <- prep_otn_tagging(tm, db = db)
validate_acoustic_animals(prepped_otn$acoustic_animals)
}
}
