% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation_fxns.R
\name{validate_table}
\alias{validate_table}
\title{Validate data for database import}
\usage{
validate_table(dat, tbl)
}
\arguments{
\item{dat}{A dataframe containing data to be appended to a SOMNI db table.}

\item{tbl}{The name of a SOMNI db table (e.g., "metadata_acoustictags" or "deployments_retrievals").}
}
\value{
Warnings if any validation checks fail
}
\description{
This function takes a supplied dataframe and a database table
to check that:
\enumerate{
\item the column names of the dataframe match the column names of the database table
\item all non-NULL columns have no NULL values
\item all column data types of the dataframe match the column data types of the database table
}
}
\examples{
tags <- prep_tag_sheet(raw_tag_dat)
validate_table(tags, "metadata_acoustictags")

otn_tagging <- prep_otn_tagging(raw_otn_data)
validate_table(otn_tagging$metadata_animals, "metadata_animals")
}
